refactor(brutus): extract usage text to usage.go
